(MOBIAPPS-2148 csa2) better fetching/retry for incidents, implement Notes piece, longs for noteIds
git-tfs-id: [http://tfs2010:8080/tfs/gd]$/Delta/Services/AtlantisFramework/4.0;C387226

+++ Triplets/Iris/Atlantis.Framework.Iris/Atlantis.Framework.Iris.Impl/Web References/irisService/Reference.cs
@@ -525,7 +525,7 @@ namespace Atlantis.Framework.Iris.Impl.irisService {
         
         /// <remarks/>
         [System.Web.Services.Protocols.SoapDocumentMethodAttribute("http://tempuri.org/GetIncidentCustomerNotes", RequestNamespace="http://tempuri.org/", ResponseNamespace="http://tempuri.org/", Use=System.Web.Services.Description.SoapBindingUse.Literal, ParameterStyle=System.Web.Services.Protocols.SoapParameterStyle.Wrapped)]
-        public string GetIncidentCustomerNotes(long incidentId, int noteId) {
+        public string GetIncidentCustomerNotes(long incidentId, long noteId) {
             object[] results = this.Invoke("GetIncidentCustomerNotes", new object[] {
                         incidentId,
                         noteId});
